+++ Geometry/Figures/General/OneDimensional.swift
@@ -6,13 +6,14 @@
 //  Copyright © 2017 Trovy. All rights reserved.
 //
 
-import Foundation
+import CoreGraphics
+import Result
 
 protocol OneDimensional: FigureBase {
 //    func intersectionPointsNotWith(other: Transmitter?) -> [Point]
     var oneDimensionalStorage: OneDimensionalStorage { get set }
     var touchingDefiningPoints: [Point] { get }
-    func at(_ pos: Float) -> RawPointResult
+    func at(_ pos: CGFloat) -> RawPointResult
     func nearest(from point: RawPoint) -> FloatResult
     func gapToCenter(from point: RawPoint) -> FloatResult
 }
